[FIX] Fixes the link to buy more credits
In EnrichmentInfo, the constructor would override the `info`
returned by the service with its own message. Then the Main would
use that message as the url to buy more credits.
Now the `info` field is not overridden anymore, and the Main uses
the `EnrichmentInfo.getTypicalMessage` function to get the message.