[FIX] account@saas~12.3: Ignore custom tax formulas
During upgrade to `saas~12.3`, some temporary invoices are created. If
there is a tax with a custom formula that uses a field that is not
present during the upgrade (because it is part of a custom module), the
upgrade will fail.

To avoid this, we skip tax computations which would otherwise fail
during the creation of these invoices.

Request ID: 55962

closes odoo/upgrade#1930

Signed-off-by: Christophe Simonis <chs@odoo.com>