[FIX] account@saas~12.4: set quantity in move lines grouping invoice lines
When invoice lines are grouped together in one move line,
set on the move line the quantity from the invoice lines

The quantity field on move lines cannot really be trusted,
in addition, in the former invoice analysis,
the quantity displayed was from the invoices rather
than the move lines. So, if we want a chance to keep the
same figures in invoice analysis, we better use the quantities
from the invoice lines rather than from the move lines.

It's already what is done anyway for regular lines matched
to move lines anyway, at the end of the script,
thanks to
```sql
            UPDATE account_move_line aml
               SET quantity = invl.quantity,
                   price_unit = invl.price_unit,
                   discount = invl.discount,
                   price_subtotal = invl.price_subtotal,
                   price_total = invl.price_total
              FROM invl_aml_mapping mapping
        INNER JOIN account_invoice_line invl ON invl.id = mapping.invl_id
             WHERE mapping.aml_id = aml.id
```

opw-2329864

closes odoo/upgrade#1949

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>