[FIX] stock@0.0.0: add match_uniq for stock.warehouse
The `stock` module provides a default warehouse with id `warehouse0`,
defined in `stock_data.xml`. If the warehouse associated with this
record is removed in the database but a new warehouse is created for the
default company with the same name as the company, upgrading the stock
module will fail.

This is because when reading the `warehouse0` record from the XML file,
it will try to create a `stock.warehouse` for the default company with
the same name as the company (as this is how the default name for
`stock.warehouse` is defined), which will cause a crash.

To avoid this problem during version upgrades, we will ignore the
insertion of repeated records for this model.

Request ID: 44636

closes odoo/upgrade#1960

Signed-off-by: Christophe Simonis <chs@odoo.com>