[FIX] account@13.0: badly formatted aliases being updated
In commit
https://github.com/odoo/upgrade/commit/b5909643ccd7dfaff591359f6ac0c755ac5963c9,
mail aliases linked to a journal are updated, or else they stop working
after upgrading to `13.0`.

However, a check introduced in versions `12.0` onward in
https://github.com/odoo/odoo/commit/6c154ab5f5af48480da6b77299687d9a768cd14f,
raises an error if a mail alias is not compliant with RFC 5322, and the
check only triggers when we modify an alias, in particular, when we try
to update it. So it may be the case that a customer has a non-compliant
alias that has gone undetected, and when updating it in the upgrade
script, the check raises the error making the upgrade fail.

Since it is not the upgrade's responsibility to correct malformed
aliases, this check is completely skipped during the upgrade to avoid
having the upgrade crash when an alias is modified.

Request ID: 55750

closes odoo/upgrade#1967

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>