[IMP] helpdesk_timesheet/saas~14.2: remove task_id in helpdesk ticket
Before, we could link a helpdesk.ticket to a project.task, but this
behaviour is not really used, since we can add timesheets in a ticket
and also create a sale order.

That's why the task_id field is removed of the helpdesk.ticket.

task-2388500

COM PR odoo/odoo#64474
ENT PR odoo/enterprise#15742