[IMP] account@saas-14.1: avoid tax tag inversion on misc operations
Migration of the following PRs:
https://github.com/odoo/odoo/pull/62536
https://github.com/odoo/enterprise/pull/15031

Basically, we add a new field on account.move.line telling whether or not the balance from this line needs to be inverted before applying the tag's sign, and "reinvert" the tags of the existing move lines that got inverted on the database we want to migrate. This way, we ensure the data seen by the user are consistent before and after the migration.

Note that there IS a corner case (invoice with a negative line with some tax creating a cash basis entry) in which we won't reinvert the tag. This is due to the fact we cannot handle this case with 100% accuracy (if a credit note was reconciled with it, how do you distinguish the base lines from each of them?). Hence, we prefer not to reinvert it at all. It will not make any difference in the tax report, and can be fixed manually if the user complains about it (it'd be a purely cosmetic change).

closes odoo/upgrade#1983

Related: odoo/odoo#62536
Related: odoo/enterprise#15031
Signed-off-by: Christophe Simonis <chs@odoo.com>