[ADD]: Add integrity test for account totals
Check the total balance per account as given by the sum of move lines.

Debit/Credit may change over an upgrade but the balance must remain the same.

During the payment refactoring of saas~13.4.1.1 some moves may change state,
draft->posted, therefore we perform the integrity check only if upgrading
from 14.0 onwards.

Due to a conflicting UpgradeCase in runbot, we need to skip the current test
from 14.0 to saas~14.3 when we detect data for TestTagNoInvert on the DB.

closes odoo/upgrade#2034

Signed-off-by: Christophe Simonis <chs@odoo.com>