[FIX] account_asset@saas~12.3: specify ref in account.move.line
The `ref` field is a related field in `account.move.line` that points to
the `ref` field in `account.move`. When moves and move lines are created
from depreciation lines during the upgrade, this field is not specified
during the creation of the move lines, and since they are created
through a SQL query and not through the ORM, it is not populated.

closes odoo/upgrade#1999

Opw: 2323479
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>