[FIX] mrp/saas~13.4: use psycopg query expansion
On 'Update the MO source and create backorder related' if the list of
inserted/updated records is empty some queries fail later on. This is
due to the presence of string formatting of SQL queries on Python.
Here we change the code to rely on the safe psycopg query expansion.

closes odoo/upgrade#2027

Signed-off-by: Christophe Simonis <chs@odoo.com>