[FIX] account@saas~13.4: force a lock date on companies
The script `account/saas~13.4.1.1/post-10-payment-refactoring.py`
entirely relies on the fact the companies lock date
to not modify the accounting of past years.

Nevertheless, for small and medium companies,
it's highly likely the users did not set the lock dates.

Hence, it could potentially, for instance,
modify the accounting of 2015 for companies without lock date set.

To overcome this issue,
we force a lock date for companies without lock date set,
or older than 1.5 years.

closes odoo/upgrade#2039

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>