[IMP] helpdesk_timesheet/saas~14.2: store total_hours_spent field in ticket
We store this field to get the correct total hours spent. That is, the
sum of the unit_amount of all timesheets contained in the ticket,
because a user has "See own timesheets" group cannot see the
timesheets of another employees and then he doesn't have the correct
total hours spent for this ticket.

task-2409761

COM PR odoo/odoo#64474
ENT PR odoo/enterprise#15742