[IMP] helpdesk_timesheet/saas~14.2: timesheet can be linked either task or ticket
Before, the ticket could be linked to a task. Now, it's not the case,
then the timesheet would not be linked to task and a ticket at the same.

This commit updates the existing timesheets to set task_id to NULL if
the timesheet was linked to task and ticket at the same time.

task-2424382

COM PR odoo/odoo#64474
ENT PR odoo/enterprise#15742