[IMP] base/saas~14.2: add helpdesk_fsm_sale module
This module is used to make a bridge between Helpdesk and FSM Sale.

This module is necessary because we need to restrict the pricing type
helpdesk projects to task rate as fsm projects. However, the fsm
projects also need the timesheet_product_id field in the 'Invoice'
page of project form view, compare to helpdesk projects which do
not need it and thus we can directly hide the 'Invoice' page.

task-2424382

COM PR odoo/odoo#64474
ENT PR odoo/enterprise#15742