[FIX] stock/0.0.0: add a new method to handle uom inconsistencies
In some cases, the MOST_USED method that consists in selecting the most
used UoM category among the stock moves for each product leads to
change the UoM of some products.

Changing product's UoM may also lead to stock quantity inconsistencies.

Another solution would be to always use the UoM of the product as a reference
to upgrade all the stock moves.

This PR handles both methods according to the value of the environment
variable ODOO_MIG_ENABLE_UOM_INCONSISTENCIES_FIX.

closes odoo/upgrade#2083

Request: 56386
Signed-off-by: Christophe Simonis <chs@odoo.com>