[IMP] sale_timesheet/saas~14.2: only display one time the warning
When the following condition for a SOL:
(delivered quantity / ordered quantity) >= threshold set in the product
is True, then we display an upsell warning for the corresponding SO.
The problem is we don't check if the warning has already been displayed
by the same SOL. Thus, each time we timesheet for this SOL, we display
one more time a warning.

This commit adds a new boolean field in sale.order.line to avoid to
spam the salesman.

odoo/odoo#64252
task-2411291

closes odoo/upgrade#2066

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>