[IMP] social/saas~14.4: allow multiple companies to use Social
Purpose
=======
Allow multiple companies to use Social on the same database.

So each company does not want to share their social accounts with the
others companies.

We let the company on the social post empty, to be sure to not break
the new constrains.

Task 2415987
See odoo/enterprise/pull/15896

closes odoo/upgrade#2099

Related: odoo/enterprise#15896
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>