[ADD] {base,web}/saas~14.3: Migrate asset bundle from view to ir.asset and manifest
Converts both definitions and extension of asset bundle that were previously ir.ui.views that we could
call with `t-call-assets`

Expressivity of the new system is poorer but simpler. As a consequence several use cases are not migrable automatically.
In particular, there is not an easy way to interpret xpath, so a simple heuristic is put in place:
- we can only detect if some string matching a regex is present and compute the ir.asset's target accordingly
namely, the most used is `//link[last()]` which translates into a `append` directive for the ir.asset
When a specific node is to be targeted in the xml, the xpath needs to target its src or href attribute and nothing else.
- `t-call` that have content (that will be evaluated downstream with a `t-raw="0"`) have no equivalent and therefore can't be migrated
- dynamic python expression can't be migrated
- node attributes can't be migrated

task-id: 2352566
odoo/odoo#60632
odoo/enterprise#14668

closes odoo/upgrade#2122

Related: odoo/design-themes#454
Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Bruno Boi <boi@odoo.com>
Co-authored-by: Lucas Perais <lpe@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>
Co-authored-by: Simon Genin <ges@odoo.com>