[FIX] stock_account/saas-12.4: migrate price history
Use case:
- Create product A with cost 10$ 	(date 1 January)
- Update quantity to 10 units 		(date 1 January)
- Change cost to 20$ 			(date 2 January)
- Update quantity to 20 units		(date 3 Janaury)

12.0 Inventory Valuation at date 2 January:
- Product A 10 units 200$

13.0 Inventory Valuation at date 2 January (svl):
- Receipt 10 units at cost of 10$	100$
- Update of cost from 10$ to 20$        100$

13.0 Migrated database:
- Receipt 10 units at cost of 10$	100$

So the migrated database are missing the cost change in the past.
The value at current date is correct because at the end of layer
migration, we check if the value match the quantity * product_cost
(for standart price and avco) and add a layer with the missing value.

This PR add the change of cost layer at the current date instead of
doing a single correction at the end.