[IMP] sale_timesheet@14.3: move timesheet_product_id to industry_fsm_sale
Before this commit, the timesheet_product_id field in the
project.sale.line.employee.map model is never used in sale_timesheet.

This commit moves the field in the industry_fsm_sale module in which it
is useful when the pricing type of the fsm project is employee rate.
However, if the industry_fsm_sale is not installed, then the field is
removed.

task-2376382

Community PR: odoo/odoo#66157
Enterprise PR: odoo/enterprise#16348

closes odoo/upgrade#2160

Signed-off-by: Christophe Simonis <chs@odoo.com>