[IMP] project@14.3: add task dependencies feature
This commit adds two fields, one for in the settings of the Project App
and the other for the project configuration.
These fields are used to add a new feature called "Task Dependencies" in
the Project App and this feature allows to the user to determine the
order in which to perform tasks.
Moreover, it adds the new depend_on_ids field in project.task model.

task-2387984

Community PR: odoo/odoo#66740
Enterprise PR: odoo/enterprise#16627

closes #2203

Signed-off-by: Christophe Simonis <chs@odoo.com>