[FIX] l10n_it_edi@saas-13.5: Skipped duplicate records because the standard sql_constraint doesn't allow to insertion of non-unique records.
The record is being inserted in 'account_edi_document' according to the new version standards. This is given in the reference below.
https://github.com/odoo/upgrade/blob/508157d4d6e7ec43cc704c1f818e966c7731f172/migrations/l10n_it_edi/saas~13.5.0.3/post-migrate.py

Due to which the sql_constaraint only allows unique records to get inserted. This is given in the following reference.
https://github.com/odoo/odoo/blob/14.0/addons/account_edi/models/account_edi_document.py#L27

but due to the insertion of duplicate records in some databases, migration gets failed.

closes odoo/upgrade#2171

Signed-off-by: Christophe Simonis <chs@odoo.com>