[FIX] matt: handle a race-condition in the management of worktrees
When run by upgradeci, matt is isolated in a container with a shared
cache.
The cache contains the bare repositories that are used via local
worktrees.
The cleaning of these worktrees cannot be done blindly via
`git worktree prune` as it will also remove the references to the
worktrees in concurrent matt runs.

Most of the time, this race-condition passed under the radar as the
only git operation we do is expected to fail (but not for this reason).
It actually fails for older versions (< 12.0) that we don't test often,
hence the late discovery.

closes odoo/upgrade#2183

Signed-off-by: Christophe Simonis <chs@odoo.com>