[FIX] base@tests: do not crawl as sudo
`env = self.env.user.with_context(**context).env`
had as side-effect to make `env` becomes sudo
because `self.user.user` is browsed as sudo.

This wasn't the expected effect. The expected
effect was just to create an env with the additional context
from the action.

First, we do not want to crawl as sudo,
to check more things (such as record rules).

Second, in the case of the given request,
it leaded to performance issues,
because the user was reading way more data
then he would normally sees because of the sudo.

For instance, a group by leaded to a memory error
because the record rules were not applied
and there was more data then he should normally see.

However, when grouped by a relation field
(e.g. a many2one)
the group name must be computed using `sudo`
as in standard:
https://github.com/odoo/odoo/blob/a4d7e01e2a27e90060ffd19dffbad3cead5d608b/odoo/models.py#L1901

request 59526

closes odoo/upgrade#2188

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>