[FIX] base@tests: crawl the user company only
From Odoo 13.0, this is possible to browse
the database with multiple companies at a time.

When `allowed_company_ids` is not passed in the context,
this is what will happen by default, all the user
companies will be used,
and therefore it will get the data from all the companies
during the crawling.
When there are a large number of companies, with a large
number of data, this can lead to performance issue.

Besides, this doesn't represent how the user
is actually browsing his database. He will hardly
look all the company at the same time.

For performance reasons,
only crawl the current user company (`user.company_id`).

If we want to validate/crawl the database for all companies,
then the solution would be to crawl each company at a time,
in a loop on companies, and not all at once.

request 59526

closes odoo/upgrade#2189

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>