[FIX] website_helpdesk_form/saas~13.1.1.0: query deadlock
The call to _ensure_submit_form_view on
https://github.com/odoo/upgrade/blob/b2deb0a751d078a55f81886ca62321b753ffe534/migrations/website_helpdesk_form/saas%7E13.1.1.0/end-ensure-submit-form-view.py#L7
triggers the installation of additional modules here
https://github.com/odoo/enterprise/blob/be1ab29ba1f1aecb7370dfdd19ae7f4156c3b4d2/helpdesk/models/helpdesk.py#L178
this provokes a deadlock on a query on ir_act_server (multiple
transactions, one of which modifies that table).

Here we uncheck for modules that have been uninstalled but are still marked as
used by helpdesk teams. This ensures the installation does not happen later on
the conflicting end script.

upg-6983

closes odoo/upgrade#2219

Signed-off-by: Christophe Simonis <chs@odoo.com>