[FIX] account/saas~13.4: set value of is_move_sent of moves related to payments
Before this commit, payments that are sent or reconciled in V13 are showing up
in a filter named 'SCT Payments To Send' in V14 after migration, which is wrong.
After this commit, the field `is_move_sent` in account_move related to payments
is set to True if the payment was sent or reconciled in V13.

opw-2462681

closes odoo/upgrade#2226

Signed-off-by: Christophe Simonis <chs@odoo.com>