[FIX] account/saas~13.4.1.1: bad journal type
Some client DBs have banks statements linked to journals with the wrong
type: 'general' (not 'cash' or 'bank'). This raises an exception while creating
the draft account moves linked to those bank statements in the post payment
refactoring. Specifically the suspense_account_id of the journal is set
to False, but it is needed for the drafts
_prepare_move_line_default_vals.

Here we set the suspense_account_id of the conflicting journals to the
one set up for the respective company.

upg-8612
upg-6126
upg-2558

closes odoo/upgrade#2225

Signed-off-by: Christophe Simonis <chs@odoo.com>