[IMP] crm_iap_lead/SaaS~14.4: change lead.mining.request's error to error_type
This commit transforms the existing "error" field on the
crm.iap.lead.mining.request model, that was never actually used, to a more
versatile error_type that is used to display an error block on top of the form
view.

See community commit for more details.

COM PR odoo/odoo#67574
Task-2478427

closes odoo/upgrade#2261

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>