[ADD] gmail: create an Gmail add-on to enrich and manage lead and ticket
Purpose
=======
Create a Gmail add-on to link Odoo to Gmail. This add-on show the
related partner when we open an email, allow to manage leads (from CRM)
and tickets (from Helpdesk) from Gmail.

Specifications
==============
IAP
---
The add-on can work without a Odoo database, by contacting IAP.

Each user can make 400 free enrichment. So when they open an email, we
try to enrich the sender (of receiver if it's an email we sent).

Client database
---------------
The user can link his add-on to his Odoo database, to be able to see
his partner / company information when he open an email.

So when we open an email, we search the related partner in the user
database and the related lead and ticket.

The user can create leads and tickets from the add-on. The email body
is written on the description field, and the user is redirected to the
form view of the record to be able to edit the information of the new
record.

The user can log the email on each partners / leads / tickets.

The user can search a new partner to open in the add-on. So if we took
the wrong partner, or if the user want to log the email on a different
partner (e.g.) he can.

Technical
=========
Authentication
--------------
Here are the steps for the authentication to the user database,
1. Create a state token. The state token is used to execute a function
   outside of the add-on. So if the user is redirected to
   "/url?state=<code>" the function we gave to the state token builder
   will be executed (it's our callback endpoint)
2. Redirect to the Odoo database with the state in parameter
3. The user accept the permission in Odoo and is redirected to the
   callback URL ("/url?state=<code>&auth_code=<code>")
4. The Google server look at the state token and know which function
   to execute.
5. Our function take the "auth_code" and make a RPC call to the Odoo
   database to exchange this "auth_code" to an "access token"
6. The access token is save in the "user properties" and then a JS code
   close the tab in the browser
7. When the tab is closed, the add-on is automatically refreshed, read
   the access token from the "user properties" and can now make request
   to Odoo

Limitation
----------
To not let the user log 2 times the same email on the same partner,
lead, ticket, we store the logging information in the cache. But the
maximum life time of a cache key is 6 hours, it means that after that
period of time the user will be able to re-log the same email on the
same record.

Task 2115131
PR #3
See odoo/iap-apps/pull/307
See odoo/odoo/pull/66139
See odoo/enterprise/pull/16913