[FIX] planning/saas~13.2: assets_common_planning already renamed
In `saas-13.1`, the view was named `planning.assets_frontend_planning` and
then was renamed to `planning.assets_common_planning` in `saas-13.2`,
so it has been taken into account in a migration script.

Then, in the commit odoo/enterprise@2596704,
this view was directly named to `planning.assets_common_planning` in Odoo `13.0`, so
the migration script was not working anymore.

upg-3392