[FIX] util: fix updating domain/context
When a field is removed/renamed we need to update the domains and
specific parts of contexts where it appears. This includes dashboards
and filters.

Here we improved the cleaning of the domains, and added extra keywords
that need to be verified on contexts. We also add some tests.

The new cleaning takes an adapter function to adapt the leafs of domains
that have the updated field as last component on the left. This allows
to switch leafs of removed fields with TRUE_LEAF/FALSE_LEAF depending on
the operators being applied.

Depends on odoo/odoo#68615 and #2357

upg-10669
opw-2487322

API-BREAKING change.

closes odoo/upgrade#2320

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Christophe Simonis <chs@odoo.com>