[FIX] util: avoid usage of t-raw in migration report
See odoo/odoo#68072 for the logic behind that bit: t-raw is deprecated
and Markup is supported by qweb. Can't exactly change it to t-esc since
including HTML is the entire point, so instead if
`odoo.tools.misc.html_escape` is `markupsafe.escape` then we apply the
transformation. Otherwise, leave things as-is.

closes odoo/upgrade#2315

Related: odoo/enterprise#17228
Signed-off-by: Christophe Simonis <chs@odoo.com>