[FIX] mail@9.0: improve `mail_message` deletion performances
This cleaning script deletes mail messages massively.

For instance, in the case of the studied databases,
it deletes more than 200.000 procurement order messages.

Removing this constraint temporary speeds up the
deletion from days to a few minutes, thanks
to the triggers not being checked.

Normally, there is no risk deleting messages
with this constraint removed,
as the `parent_id` of a message is always supposed to be on the same
`res_model`/`res_id`, and the point of this script is to remove
messages on no longer existing `res_model/res_id`,
therefore deleting the entire thread, and therefore
the parents along its children.

closes odoo/upgrade#2354

Signed-off-by: Christophe Simonis <chs@odoo.com>