[IMP] base/0.0.0: add mechanism to ignore unlink of some xmlids
As for views/fields/modules/models sometimes a data can be
considered deleted when added in stable and not yet forwardported.

This commit adds an mechanlism to define specific xmlids to ignore,
mainly used by runbot during the forward port of pull requests
adding data in stable, like odoo/odoo#68785.

closes odoo/upgrade#2355

Signed-off-by: Christophe Simonis <chs@odoo.com>