[FIX] base/saas~12.3: fix install sale prod conf
If 'Variants' is not checked in Settings and we install
sale_product_configurator we get a warning on save in Settings.
This happens because the setting 'Product Configurator' is not checked
when 'Variants' is unchecked.

To solve this sale_product_configurator should be installed only
when 'Product Variant' is checked on Settings.

As a workaround to check the settings on a pre- script we check the
implied groups of base.res_group, since checking 'Variants' will add
product.group_product_variant to the implied groups.

upg-10669
opw-2487336

closes odoo/upgrade#2363

Signed-off-by: Christophe Simonis <chs@odoo.com>