[REF] {base,website}/saas~14.4: rename assets 'glob' to 'path'
This change was meant to be merged in 14.3 before the freeze, but it
came too late and now it's a pain to get it merged.

Trying to do it in master first, and then to backport it in 14.3.

Seems to require some ugly changes in the 14.3 scripts too, for the
one-pass upgrade between 14.0 and master.

See odoo/odoo#68695
See odoo/enterprise#17465
See odoo/design-themes#459

closes odoo/upgrade#2348

Signed-off-by: Olivier Dony (odo) <odo@openerp.com>