[IMP] sale_timesheet/14.4: simplify pricing_type in project
- Remove column of the pricing_type

The column for pricing_type field in the
project_project table is removed because this field will be
a compute non-stored. This change needs to check if the pricing_type
column exists in project_project table in industry_fsm_sale module
before updating in 14.2.

- Remove link_selection field in create SO wizard

This field is no longer used in this wizard because this field is
invisible and in this wizard, the user can only create a new SO.
Thus, this commit removes this unused field.

- Remove pricing_type field in create SO wizard

This field was a related field to the one in project.
However, the picing_type is the project has been changed and
now it is a compute non-stored field. Thus, the condition made in the
wizard must be changed to not used this field. For example, if the
pricing_type of a project is fixed then it means this project has a
SO and SOL and thus this wizard will not create a new one for this
project.

- Remove sale_line_id in wizard lines

The sale_line_id field in project.create.sale.order.line model is
used when the user wants to link an employee mapping to SOL on
existing SO with the project.create.sale.order wizard.
But since the link_selection field has been removed, this behaviour
is no longer available and used. So the sale_line_id is useless in
this model. That is why this field is removed of this model.

- Convert sale_order_id in project into related field

This commit removes the column of the sale_order_id field in the
project_project table, because this field is no longer stored, it will
be a related field. The reason is because this field become useless to
edit in the project form view. The default sales order item in the
project is no longer filtered to have only the SOL in the SO but now the
user can select any SOLs for the customer defined in the project. Same
thing for the SOL in task, timesheets and employee mapping.
That's why this field is converted into a related field.

- Remove the constaint called `sale_order_required_if_sale_line` in
project.project model.

This constraint is used to check if the sales order item is defined
when there is a sales order in the project.
But since the sales order is now a related this constraint is useless.

task-2442683

Community PR: odoo/odoo#66181
Enterprise PR: odoo/enterprise#16356

closes #2164

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>