[IMP] outlook: add a fallback image company contacts
Add the image used in the odoo kanban view if:
- The company has no image in the database
- The image provided by the enrichment url is corrupt

Purpose is to make the image displayed in the plugin consistent with the one
displayed on the Odoo contacts app so that the user does not get confused.

Make use of extra space for phone numbers so that phone numbers don't take
more than one line.

Avoid redundancies in company insight fields when the parent partner is a
company, as it does not make sense to show the same fields twice in the profile
card and in the insights section.

Task-2496443
Community PR odoo/odoo#69175
Client PR odoo/mail-client-extensions#7