[FIX] l10n_us_1099/saas~14.3: mute logger
The demo data of this new auto-install module cannot be installed due to
accounting lock dates set by the `account/tests/test_lockdate.py`
upgrade test.
This generate a traceback that is considered as an error by runbot.

This demo data is not required and this warning can be ignored.