[IMP] account@saas~13.4: speeds up payment deletion
Creating the index, at the right moment
as soon as possible before the deletion so the index is ready to be used by postgres,
is actually the critical part to make things really faster.

Using raw sql would speed up things even more because:

- The ORM, when making the list of `mail.message` to delete,
  prefetch all fields of mail.message rather than just the list of the message ids.
  And if there is a lot of payment to unlink, it prefectches columns for a lot of messages.
- a raw `DELETE` woudln't rely on a IN operator with a huge ids list,
  which is less performant because it creates a lot of locks in postgres.

The unlink with a lot of ids is actually what is causing the LWLockRelease issue I am investigating for days:
Postgres will create one or two locks by rows to delete. Besides, the ORM delete by chunk of 1000 records:

- https://github.com/odoo/odoo/blob/070c2853956c97bd64aac6abe0e208ef11ef8d3d/odoo/models.py#L3442
- https://github.com/odoo/odoo/blob/070c2853956c97bd64aac6abe0e208ef11ef8d3d/odoo/sql_db.py#L331²
- https://github.com/odoo/odoo/blob/070c2853956c97bd64aac6abe0e208ef11ef8d3d/odoo/sql_db.py#L224

And it therefore explains why the number of locks caps around 2000 locks when I monitored pg_locks.

I therefore come to the conclusion that deleting records with a huge IN ids list is really bad,
not only for the database itself but for the entire cluster.

In this case, it's even more amplified by the fact the DELETE was slow because of this missing index.

> Wouldn't it be better to call util.remove_records, which cleanup linked records?

remove_records works with an xmlid (or a lone tuple (model, res_id)), so it's doesn't apply here
I see I could use the low level _remove_records instead but
it's a low-level method so I guess it's not intended to be used in an upgrade script.
For instance it's not used in any upgrade script for now.
Besides, it doesn't really solve the IN with a huge list of ids.

> You can also maybe call unlink through util.iter_browse()

I can maybe use it to reduce the unlink batch number from 1000 to something lower to decrease the LWLockrelease effect stated above.
(So there is less locks by chunks)
But what I am worried is that we should apply that strategy to any unlink
processing a huge list of ids so maybe there is something to do with unlink more generically.
For the moment maybe I can apply iter_browse for this specific case only,
and if in the future we see other cases of unlink with a large number of ids, we can implement a generic solution.

```
EXPLAIN ANALYZE
DELETE
  FROM account_payment
 WHERE id IN (
    SELECT pay.id
      FROM account_payment pay
      JOIN account_payment_pre_backup pay_backup ON pay_backup.id = pay.id
     WHERE pay_backup.state NOT IN ('draft', 'cancelled')
       AND pay.move_id IS NULL
     LIMIT 10
     )
                                                                                  QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Delete on account_payment  (cost=20.37..44.45 rows=10 width=34) (actual time=0.449..0.450 rows=0 loops=1)
   ->  Nested Loop  (cost=20.37..44.45 rows=10 width=34) (actual time=0.383..0.404 rows=10 loops=1)
         ->  HashAggregate  (cost=19.95..20.05 rows=10 width=32) (actual time=0.378..0.380 rows=10 loops=1)
               Group Key: "ANY_subquery".id
               Batches: 1  Memory Usage: 24kB
               ->  Subquery Scan on "ANY_subquery"  (cost=0.42..19.92 rows=10 width=32) (actual time=0.035..0.371 rows=10 loops=1)
                     ->  Limit  (cost=0.42..19.82 rows=10 width=4) (actual time=0.032..0.365 rows=10 loops=1)
                           ->  Nested Loop  (cost=0.42..186848.76 rows=96312 width=4) (actual time=0.031..0.364 rows=10 loops=1)
                                 ->  Seq Scan on account_payment_pre_backup pay_backup  (cost=0.00..14763.48 rows=277317 width=4) (actual time=0.015..0.086 rows=53 loops=1)
                                       Filter: ((state)::text <> ALL ('{draft,cancelled}'::text[]))
                                       Rows Removed by Filter: 166
                                 ->  Index Scan using account_payment_pkey on account_payment pay  (cost=0.42..0.62 rows=1 width=4) (actual time=0.005..0.005 rows=0 loops=53)
                                       Index Cond: (id = pay_backup.id)
                                       Filter: (move_id IS NULL)
                                       Rows Removed by Filter: 1
         ->  Index Scan using account_payment_pkey on account_payment  (cost=0.42..2.44 rows=1 width=10) (actual time=0.002..0.002 rows=1 loops=10)
               Index Cond: (id = "ANY_subquery".id)
 Planning Time: 0.294 ms
 Trigger for constraint account_move_line_payment_id_fkey: time=1.236 calls=10
 Trigger for constraint account_payment_sale_order_rel_account_payment_id_fkey: time=0.195 calls=10
 Trigger for constraint credit_token_payment_id_fkey: time=1.308 calls=10
 Trigger for constraint payment_transaction_payment_id_fkey: time=0.288 calls=10
 Trigger for constraint account_invoice_payment_rel_payment_id_fkey1: time=0.191 calls=10
 Trigger for constraint account_move_payment_id_fkey: time=6015.799 calls=10
 Trigger for constraint account_payment_account_bank_statement__account_payment_id_fkey: time=0.444 calls=10
 Execution Time: 6019.988 ms
(26 rows)
```

upg-12139

closes odoo/upgrade#2456

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>