[FIX] base/10.saas~18: allow to force procurement uninstall
Before Odoo `11.0`, the `procurement` module is auto-installed when
the `sale` app is installed, and a `procurement.order` record is
automatically created each time a `sale.order` is created.

Then, in Odoo `11.0` the `procurement` module is merged with
the `stock` module, leading to auto-install the `stock` app.

Unfortunately, in some cases, customers just want to create
sale orders without any procurement orders, and then don't want
to automatically get the stock app installed in the upgraded database.

This PR adds a new environment variable `ODOO_MIG_FORCE_PROCUREMENT_UNINSTALL_11`
to allow forcing the uninstall of the `procurement` module.

upg-60137

closes odoo/upgrade#2470

Signed-off-by: Christophe Simonis <chs@odoo.com>